bip32

Trait PrivateKey

pub trait PrivateKey: Sized {
    type PublicKey: PublicKey;

    // Required methods
    fn from_bytes(bytes: &PrivateKeyBytes) -> Result<Self>;
    fn to_bytes(&self) -> PrivateKeyBytes;
    fn derive_child(&self, other: PrivateKeyBytes) -> Result<Self>;
    fn public_key(&self) -> Self::PublicKey;

    // Provided method
    fn derive_tweak(
        &self,
        chain_code: &ChainCode,
        child_number: ChildNumber,
    ) -> Result<(PrivateKeyBytes, ChainCode)> { ... }
}

Trait for key types which can be derived using BIP32.

Required Associated Types

type PublicKey: PublicKey

Public key type which corresponds to this private key.

Required Methods

fn from_bytes(bytes: &PrivateKeyBytes) -> Result<Self>

Initialize this key from bytes.

fn to_bytes(&self) -> PrivateKeyBytes

Serialize this key as bytes.

fn derive_child(&self, other: PrivateKeyBytes) -> Result<Self>

Derive a child key from a parent key and the a provided tweak value, i.e. where other is referred to as “I sub L” in BIP32 and sourced from the left half of the HMAC-SHA-512 output.

fn public_key(&self) -> Self::PublicKey

Get the Self::PublicKey that corresponds to this private key.

Provided Methods

fn derive_tweak( &self, chain_code: &ChainCode, child_number: ChildNumber, ) -> Result<(PrivateKeyBytes, ChainCode)>

Derive a tweak value that can be used to generate the child key (see [derive_child]).

The chain_code is either a newly initialized one, or one obtained from the previous invocation of derive_tweak() (for a multi-level derivation).

Warning: make sure that if you are creating a new chain_code, you are doing so in a cryptographically safe way. Normally this would be done according to BIP-39 (within [ExtendedPrivateKey::new]).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.