bip32

Trait PublicKey

pub trait PublicKey: Sized {
    // Required methods
    fn from_bytes(bytes: PublicKeyBytes) -> Result<Self>;
    fn to_bytes(&self) -> PublicKeyBytes;
    fn derive_child(&self, other: PrivateKeyBytes) -> Result<Self>;

    // Provided methods
    fn fingerprint(&self) -> KeyFingerprint { ... }
    fn derive_tweak(
        &self,
        chain_code: &ChainCode,
        child_number: ChildNumber,
    ) -> Result<(PrivateKeyBytes, ChainCode)> { ... }
}

Trait for key types which can be derived using BIP32.

Required Methods

fn from_bytes(bytes: PublicKeyBytes) -> Result<Self>

Initialize this key from bytes.

fn to_bytes(&self) -> PublicKeyBytes

Serialize this key as bytes.

fn derive_child(&self, other: PrivateKeyBytes) -> Result<Self>

Derive a child key from a parent key and a provided tweak value.

Provided Methods

fn fingerprint(&self) -> KeyFingerprint

Compute a 4-byte key fingerprint for this public key.

Default implementation uses RIPEMD160(SHA256(public_key)).

fn derive_tweak( &self, chain_code: &ChainCode, child_number: ChildNumber, ) -> Result<(PrivateKeyBytes, ChainCode)>

Derive a tweak value that can be used to generate the child key (see [derive_child]).

The chain_code is either a newly initialized one, or one obtained from the previous invocation of derive_tweak() (for a multi-level derivation).

Warning: make sure that if you are creating a new chain_code, you are doing so in a cryptographically safe way. Normally this would be done according to BIP-39 (within [ExtendedPrivateKey::new]).

Note: child_number cannot be a hardened one (will result in an error).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.